Added the service collection extensions and automapper

+++ NotifyMe.Core/Entities/AlertTrigger.cs
@@ -3,7 +3,7 @@
 public class AlertTrigger: BaseEntity
 {
     public string? AlertName {get;set;}
-    public long EventMonitoringId {get;set;}
+    public int EventMonitoringId {get;set;}
     public virtual EventMonitoring? EventMonitoring {get;set;}
     public string? ThresholdCondition {get;set;}
 }
